Fold the two main-thread routes into one
Review of the branch turned up a regression it had introduced: closing the
wrapper's editor posted `close_all_editors` and then dropped the queue it
had just posted to, whose `Drop` clears it. On Windows — the platform that
worked — sub-plugin windows survived the editor closing, and any command
from the last frame went with them. On `main` this was a synchronous call.

The cause was structural rather than local: there were two ways to reach
the main thread, `host-window`'s `Deferred` and `Shared::post_main`, and
the editor chose between them. `Deferred` had one consumer in the
workspace, was never constructed on X11, and belonged to the editor —
which is what let an editor cancel its own work. `post_main` does the same
job on every platform and its queue belongs to the plugin instance, so it
is now the only route and the trap is unrepresentable.

That deletes `Deferred` in all three backends, the queue plumbing it
needed on the X connection, and the branch in the editor. `poll` no longer
runs any caller's code at all — an event only writes to a `WindowState` —
so there is no longer a turn of it that is unsafe to be inside of.

Also from the review, in descending order of how much it would have hurt:

- `FrameImpl` kept raw pointers into a `PlugFrame` it does not own. A
  plugin holds `IRunLoop` until its own teardown, which is after the host
  has let the frame go, and the wrapper is `Arc`-backed — so unregistering
  wrote into freed memory. The state lives in the COM object now, which is
  the thing whose lifetime the plugin actually extends. There is a test
  that drops the handle and then calls in.
- `tick::run` returned before `reclaim()` when the callback landed off the
  main thread, which on Linux VST3 with the editor closed is every time.
  Programs handed back by the audio thread were never freed. `reclaim` is
  a swap and a drop, so it moves above the check; the skip now says so
  once instead of being silent.
- The same "poll a list, call back, re-check liveness" loop existed four
  times with three different rules — CLAP's timers did not re-check, so a
  timer unregistered by an earlier callback still fired. One
  `host_window::watch`, used by both backends, tested once.
- `publish_graph` held the patch lock across `compile`, which on X11 is
  the editor's next frame; and took it three separate times, so the
  recorded error could describe a graph that had since changed.
- `shared.rs` imported `View` from `editor.rs`, pointing the state at the
  UI. `View` has its own module and builds from the sub-plugin host alone.
- Forwarding a key read the whole keyboard mapping from the server; it is
  cached and dropped on `MappingNotify`.
- `fault.rs` was gated on unix-not-macOS but is glibc and Linux in fact.
- Whether a sub-plugin's window is open sat behind `generation`, which
  only moves when a command runs — so a window the user closed from its
  own title bar left the node's button lit for good.
- A dead `View` field, a timer period that could overflow `Instant`, a
  duplicate `libc` dependency, `ContainerWindow::pump_events` ignoring its
  own `self`, and a lock-order note describing a hazard that does not
  exist while saying nothing about the one that does.

+++ crates/host-window/src/x11/window.rs
@@ -82,7 +82,7 @@ impl Window {
             .map_err(|e| format!("could not create the container window: {e}"))?;
 
         set_title(&conn, id, title);
-        declare(&conn, id, size, owner as XWindow);
+        declare(&conn, id, size, owner as usize as XWindow);
 
         // A round trip, not just a flush: the plugin is about to be handed this
         // id and will reparent into it from a connection of its own, which can
